fix: Reinstate Jetpack assets to editor assets endpoint (#44274)
* fix: Reinstate Jetpack assets to editor assets endpoint

Recent changes fixed de-registering of assets not present on the allow
list. However, it was overlooked that certain Jetpack assets were then
erroneously excluded.

Namely, the `jetpack-blocks-editor` assets were excluded as their
`jetpack-blocks-assets-base-url` dependency was also excluded. The
latter was excluded as it is an "inline" script without a `src`
attribute. The lack of a `src` meant the path-based exclusion
erroneously de-registered the script.

Using a handle-based exclusion means that both inline and external
scripts are correct assessed by the plugin filtering logic.

* changelog

* fix: Ensure WPCOM-specific Gutenberg plugin paths are allowed

WPCOM loads Gutenberg from a custom directory.

* test: Co-locate mocks with relevant test cases

* refactor: Remove unnecessary static function designation

The designation was erroneous to start.

* docs: Clarify intent for each allowed prefix handle

Avoid confusion regarding the purpose of each allowed handle. Namely,
the usage of `wp-` which is used by core WordPress scripts, but also
some WPCOM and Jetpack scripts too.