ci: Use GH token in update-stubs action
We've seen failures in this job the past two days. I suspect it's
the one unauthenticated GitHub API call hitting the very low rate limit.
Let's switch to using a (generic Actions) token.

And while we're messing with the file anyway, let's switch the checkout
from the generic Matticbot token to the special one we have to use later
for the `--reviewer` flag for consistency.