Open Graph: add new fallback image
Fixes CML-608

This leverages the Social Image Generator library to dynamically generate a better social preview image for sites, based off any image we can find from the site (using the same logic as before), but framing that image and display the site title so the whole preview looks better.

This is currently not something that can be customized via a UI yet.

- It uses the `edge` template from Social Media Generator because it seems to be the template that best fits social previews.
- The `jetpack_og_fallback_social_image_template` filter allows one to define their own template instead.
- If the Social Image Generator features are not available, we fall back to the previous behavior.
- The dynamically generated image src is stored in a transient for a day, to avoid putting too much strain on the service.