packages/search: Replace remaining uses of lodash (#44320)
Search code tends to import individual modules from lodash, such as
`lodash/debounce`, rather than importing the whole thing. This avoids an
extracted dependency on all of lodash, at the cost of parts of lodash
being included in the bundle.

We can save some bundle size by replacing the remaining uses of lodash:

* `pick( obj, array )` => `Object.fromEntries( Object.entries( obj ).filter( ( [ k ] ) => array.includes( k ) ) )`
* `pickBy( obj, predicate )` => `Object.fromEntries( Object.entries( obj ).filter( predicate ) )`
* `uniqueId()` => `let counter = 0` at the top of the file, then
  `++counter` where needed.
* `debounce( func, t )` => Same from the `debounce` NPM package.

This saves about 2K (on disk) in customberg, 11K in dashboard, and 3K in
instant-search.