jetpack: Replace use of lodash in Mailchimp block view.js (#44354)
This block, instead of using the full lodash, was importing
`lodash/debounce` directly (and therefore embedding it). Since we're
already bringing in the `debounce` package, use it here too.

Also, remove a useless `removeEventListener()` call. It never worked
because it's not passing the same function.

The savings of 1841 bytes (uncompressed) isn't much, but it's something
and it sets a better example.