feat: Editor assets endpoint allows the VideoPress block type (#44616)
* feat: Allow `videopress/video` block type

* feat: Preserve plugin assets registered within the `init` action

The `init` action is often used for plugin asset registration. We cannot
invoke the `init` action, as it often leads to fatal errors due to
duplicate constants. Instead, we preserve assets from the allowed
plugins and reinstate them in our `WP_Scripts` and `WP_Styles`
instances.

* test: Assert editor assets endpoint `init` action support

* changelog

* feat: Register full VideoPress block for REST API requests

Enable the GutenbergKit mobile editor access to VideoPress editor state,
which determines whether the block is activated.

* changelog