Writing Settings: Add Mailchimp connection (#44999)
Adds the ability to manage the Mailchimp connection within Settings -> Writing, effectively removing the necessity of pointing users to Calypso.

This connection is a bit more complex than the ones we ported so far since, apart from connecting/disconnecting, we need to also support signups and the the selection of the Mailchimp audience (aka list) that visitors will subscribe to. 

Therefore, we implemented some infrastructure changes:
- The external connections logic can receive signup links now
- The block extensions now supports admin scripts that can be loaded in admin contexts.
- The settings of the external connections package can be extended with extra settings.
- The external connections package can enqueue extra scripts for a given service.

There are a few notable differences when comparing the UIs between Calypso and WP Admin:
- The list/audience selector is named as "Audience" in WP Admin since that's the current terminology used by Mailchimp
- No notice displayed in WP Admin if there are no audiences created in Mailchimp, since the notice was a bit confusing and there will always an audience created by default (as long as you complete the Mailchimp onboarding).
- No notice displayed in WP Admin if no audience is selected, since the notice was a bit confusing, and I think it's obvious that no audience has been selected.