Rename temp feedback post status
I've renamed the post status we're using to skip the feedback inbox from
`jetpack-temp-feedback` to `jp-temp-feedback` because we were hitting
the character limit for post status (20 characters) and the posts
weren't saved because of that. Unit tests didn't pick this issue because
we use WorDBless to mock post insertion and this logic is missing.