forms: Fix flakey E2E tests (#45244)
Since the form submission is asynchronous, it can sometimes take time for the form submission to complete and thus in that case, asserting the success immediately after clicking the submit button will result in test failure. This commit ensures that we wait for the submission to complete before the assertion.