Forms: add email notification toggle (#45230)
* Add form submission email notifications toggle

I've added a toggle to control form submissions email notifications.

* changelog

* Fix failing forms unit test

* Bacwkard compability w/ grunion_should_send_email

In order to make the email notification toggle backward compatible with
the `grunion_should_send_email` filter, we have to make it take
precedence over the UI toggle. If the filter returns:
 - true: Send email regardless of emailNotifications setting
 - false: Don't send email regardless of emailNotifications setting
 - null: Use emailNotifications attribute to determine (default
   behavior)

* Add unit tests for email notifications toggle

* Update email toggle related copies

I've renamed the panel title from `Email connection` to `Email
responses` and the copy in the toggle now says `Send responses to email`