Tiled gallery: Improve image placeholder animation performance (#45344)
Change a loading animation to be applied conditionally to images as they load in the editor.

Before this change, the background-color animations always ran, indefinitely, in the editor.

This is an important performance improvement because the animations continued to run and cause significant browser rendering work, even though the background and animations may not be visible.