Fix: MapKit map.addItems validation error
Starting with MapKit JS versions 5.80.*, Map Blocks with markers create
errors that crash JS with the following error in the console:

Error: [MapKit] Map.addItems expected an Annotation, Overlay, ItemCollection, or Array of valid items at index 0, but got `[object Object]` instead.

This problem is due to a change in how 5.80.* validates passed
parameters to the addItems and showItems functions. The validation does
a "instanceof Array" check, which fails to properly recognize Array
values created in a realm different from the realm the MapKit library is
executing in.

The fix uses the currentWindow object to remake the annotations array
in the same realm as the realm that the MapKit library runs in.