Forms: Fix terms consent label colon issue (#45509)
* Remove period from labels on submission

The Terms consent block has a period in its label and when we render the
form submission we add a colon at the end which then renders as `.:`.
I'm trimming the period at the end before inserting the colon.`

* changelog

* Fix Terms consent period colon issue frontend side

Fix it also in the frontend for the ajax success response case.