Forms: merge email and push notifications settings panel (#45548)
* Form notification settings copy changes

* Merge email and push notification settings

I merged the forms email and push notification settings panels. I've
added an external link to the notifications documentation and adjusted
the copy and some margins.

* changelog

* Refactor feature flag check

I've moved the push notifications flag check to the panel subsection so
that it doesn't hide the email notifications when the flag is disabled.

* Support notifications link rendering logic

Show Jetpack support docs for self-hosted sites and WordPress.com
support docs for WordPress.com sites.

* Update tests