ci: Retry WP.org downloads in test-plugin-update
WordPress.org's plugin downloader seems to occasionally not respond to
connections, whether as a form of rate limiting or something else I
don't know.

Anyway, before failing the CI run, let's retry on transient-seeming
failures. The setting is to retry twice, with a delay of 30–37 seconds
between each try, to allow the network to calm or the rate limiting to
cool off.