codesniffer: Add `Jetpack.PHPUnit.FunctionCoversBackslash` sniff
PHPUnit 12 complains if a `@covers` annotation for a function or a
`#[CoversFunction()]` attribute has a leading backslash. Add a sniff to
catch and fix this.

Also, adjust the `CoverageHandler` for `AttributesSniff` to not produce
annotations or attributes with a leading backslash, to avoid the two
sniffs potentially getting into a loop.