pnpm: Hack packages to flag ESM files correctly as ESM
We do this by injecting a stub `package.json` into the directory holding
the ESM `.js` files, because the resolution algorithm looks at the
closest package.json rather than the package-as-a-whole's package.json.

On the down side, we have to make sure to copy the `sideEffects`
property into the stub, because that too looks at the closest one.
Hopefully there aren't more things like that.

In addition to the Gutenberg packages, the `uuid` package also has this
issue showing up now for some reason.