webpack: Set `resolve.fullySpecified` to false for Gutenberg packages
When the file is correctly flagged as containing ESM, Webpack by default
requires the full file import rather than a directory or extensionless
import. We can change that behavior by setting `resolve.fullySpecified`.

Gutenberg itself doesn't run into this because they're relying on
webpack/babel to be lenient with files that are flagged as CommonJS but
contain ESM code.