Fix typechecks
We can't deep-import the types directly anymore, so we have to start
with what we can import and use utility functions and accessors to drill
down to the types we need.

I note that `type ButtonProps = ComponentProps< typeof Button >` is not
the same as
`import type { ButtonProps } from '@wordpress/components/build-types/button/types'`
because it also includes `DeprecatedButtonProps`, but
nothing we have access to has `ButtonProps` without `DeprecatedButtonProps`
or vice versa. 🤷