Clean up most uses of `esc_js()`
`esc_js()` is a legacy of the pre-PHP-5.2 days, when PHP didn't have
`json_encode()`. It's basically a version of `esc_attr()` that handles
`'` differently to work right for code like
```html
<span onclick="doSomething( '<?php eccho esc_js( $string )?>' )">
```
(and it also randomly strips backslashes, yay).

Unfortunately its name tends to fool people into thinking it's
appropriate for "escaping JS" in other contexts; uses inside `<script>`
tags, for example, probably don't really intend for HTML entities to be
encoded inside the JS strings.

In modern WordPress, it's generally better to use `wp_json_encode()` to
produce a properly encoded string (or other object), and to use
`esc_attr()` when HTML entity encoding is needed. Before we can
introduce a phpcs sniff to push for that, though, we need to clean up
various existing uses.

General cleanups:

* Removed when passed into `wp_localize_script`'s `$l10n` parameter.
  That, for presumably hysterical raisins, does `html_entity_decode` on
  the `$l10n` parameter values, making it a no-op beyond screwing up
  `'`, backslash, and a few other characters.
* Replaced with `json_encode` when used as a jQuery selector. HTML
  entities are clearly wrong there, and other characters messed with are
  unlikely.
* Replaced with `json_encode` when wrapping a `__()` where the text
  doesn't have any HTML in it, as it's unlikely that it's expected that
  translations would introduce HTML that would need encoding (or that
  backslashes will be stripped).
* Replaced with `json_encode` when apparently passed a nonce or url, as
  those shouldn't need HTML entity encoding (or backslash stripping)
  either.
* Replaced with `intval` in a few places where it seems an integer is
  expected (e.g. `sprintf` with `%d`).

Specific cleanups:

* One place in woocommerce-analytics was weirdly manually building a
  JSON object, with one field getting `esc_js`-ed. Decided that field
  was probably safe and turned the whole thing into a single
  `json_encode` call.
* Calendly block had a few that should have been `esc_attr` (or maybe
  `esc_url`) in the first place. Apparently at some point in the past
  they were sprintf-ed into an event handler, then later changed to
  normal hrefs.
* Publicize had some weird thing turning printf format strings into JS
  concatenation. Rewrote that to do the same but with `json_encode`. 🤷
* Sharedaddy had some code copying GET values into form fields via JS,
  that seemed like they probably shouldn't be getting HTML entity
  encoded either.
* Videopress had a guid that seems unlikely to need HTML entity encoding
  either, particularly since various places seem to dump it into a URL
  without even percent-encoding it. 🙄

That leaves some infinite-scroll settings, a bunch of wordads stuff, a
bunch of tracking stuff (Google Analytics and wpcom), and some CRM files
already in `tools/phpcs-excludelist.json`. When introducing the phpcs
sniff to check for future use of `esc_js()`, these will be excluded.