cli: Improve import error output (#46269)
To be more user-friendly, we catch some errors that may indicate "you
have to run `pnpm install`" and display helpful information to the user.
But we're still dumping the whole error object, which is generally not
terribly useful because all the stack trace refers to "files" like
`node:internal/modules/esm/loader`.

Instead, when it's one of the errors we recognize, print only the
message without the trace, and do that in a dimmer color if possible.

And also fix the bolding of our message, which was probably broken
by #40786.