codesniffer: Add `Jetpack.Functions.EscJs` sniff (#46243)
WordPress's `esc_js()` function is a legacy of pre-PHP-5.2 days which
does some odd things to the input. In modern code, it will be better to
use `esc_attr()` and/or `wp_json_encode()`.

Now that most existing uses are cleaned up, add a sniff to discourage
future use. A few files with remaining uses are added to
`tools/phpcs-excludelist.json`.