Jetpack settings: override module list & debugger page titles (#46283)
* Jetpack settings: override module list & debugger page titles

Fixes JETPACK-1008

When registering the Module list page and the Jetpack Debugger page, we do not want the page to appear in the admin menu. This causes us to run into this issue:
- https://core.trac.wordpress.org/ticket/57579
- https://github.com/Automattic/jetpack/issues/46214

```
Deprecated: strip_tags(): Passing null to parameter #1 ($string) of type string is deprecated in …/wp-admin/admin-header.php on line 41
```

Let's override the title when we're on those pages to avoid the deprecation warning and have a proper page title on those 2 pages.

Co-authored-by: Weston Ruter <weston@xwp.co>

* changelog

* Change hook registration to match static declaration

See https://github.com/Automattic/jetpack/pull/46283#discussion_r2613950300

---------

Co-authored-by: Weston Ruter <weston@xwp.co>