codesniffer, phpcs-filter: Relicense as MIT
To allow for our sniffs and other code to be adapted and included in the
PHPCSStandards project, and also by WordPress-Coding-Standards, we are
relicensing the codesniffer and phpcs-filter packages under the MIT
license.

This license should be directly compatible with the BSD 3-clause license
used by PHPCS itself, and with the LGPL v3 used by some of the
PHPCSStandards projects.

Internal reference: p2y3YZ-9uM-p2